% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/higher_order.R
\name{add_higher_order}
\alias{add_higher_order}
\title{Add higher order taxonomy}
\usage{
add_higher_order(lookup = plant_lookup(), order_column = "order")
}
\arguments{
\item{lookup}{A lookup table (by default \code{plant_lookup()})}

\item{order_column}{The column within \code{lookup} that
corresponds to taxonomic order.}
}
\description{
Augment a lookup table with plant higher order data, from
http://datadryad.org/resource/doi:10.5061/dryad.63q27
}

