% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plant_lookup.R
\name{lookup_table}
\alias{lookup_table}
\title{Build a lookup table for a set of species}
\usage{
lookup_table(species_list, lookup_table = NULL, genus_column = "genus",
  missing_action = c("drop", "NA", "error"), by_species = FALSE)
}
\arguments{
\item{species_list}{Character vector of species names}

\item{lookup_table}{A lookup table, or by default \code{\link{plant_lookup}}}

\item{genus_column}{The column within \code{lookup_table} that
corresponds to genus.  By default this is \code{"genus"}, which is
the correct name for \code{\link{plant_lookup}}.}

\item{missing_action}{How to behave when there are genera in the
\code{species_list} that are not found in the lookup table.
\code{"drop"} (the default) generates a table without these
genera, \code{"NA"} will leave the non-genus taxonomic levels as
missing values and \code{error} will throw an error.}

\item{by_species}{If \code{TRUE}, then return a larger data frame
with one row per species and with species as row names.}
}
\description{
Build a lookup table for a set of species
}

