% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/higher_order.R
\name{add_higher_order}
\alias{add_higher_order}
\title{Add higher order taxonomy}
\usage{
add_higher_order(lookup = plant_lookup(include_counts = TRUE),
  order_column = "order")
}
\arguments{
\item{lookup}{A lookup table (by default \code{plant_lookup()})}

\item{order_column}{The column within \code{lookup} that
corresponds to taxonomic order.}
}
\description{
Augment the genus-family-order lookup table with plant higher order data, from
http://datadryad.org/resource/doi:10.5061/dryad.63q27
}
\details{
Data for the higher-level taxonomy lookup (http://datadryad.org/resource/doi:10.5061/dryad.63q27.2/1.1) compiled by Dave Tank and colleagues

The format of the lookup is that higher nodes are columns within the data.frame.
If a genus is a descendent of a particular node, the row for that genus repeats
the node name.  If a genus is not a descendent of that particular node, the cell is left blank.
}
\examples{
##' # see the format of the resource
head(add_higher_order())
#
# load the data.frame into memory
ho<-add_higher_order()
#
# return Rosid orders
unique(ho$order[ho$Rosidae=="Rosidae"])
#
# find the number of Conifer species in the world
sum(ho$number.of.species[ho$Coniferae=="Coniferae"])
}

