% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plant_lookup.R
\name{plant_lookup}
\alias{plant_lookup}
\alias{plant_lookup_version_current}
\alias{plant_lookup_versions}
\title{Plant lookup table}
\usage{
plant_lookup(version = plant_lookup_version_current())

plant_lookup_versions(all = FALSE)

plant_lookup_version_current(all = FALSE)
}
\arguments{
\item{version}{Version number.  The default will load the most
recent version on your computer or the most recent version known
to the package if you have never downloaded the data before.}

\item{all}{test against all version known to the package?}
}
\description{
Lookup table relating plant genera, families and orders.  Data
persists across package installations.
}

