% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plant_lookup.R
\docType{data}
\name{plant_lookup}
\alias{plant_lookup}
\title{Lookup table relating plant genera, families and orders}
\format{A data.frame with three columns and many rows}
\usage{
data(plant_lookup)
}
\description{
Lookup table relating plant genera, families and orders
}
\keyword{datasets}

